Float -> BigDecimal, handle rate input

+++ app/src/main/java/com/android/example/currencyconverter/model/entity/CurrencyResponse.kt
@@ -1,8 +1,9 @@
 package com.android.example.currencyconverter.model.entity
 
+import java.math.BigDecimal
 import java.util.*
 
 data class CurrencyResponse(
         val base: String = "",
         val date: String = "",
-        val rates: TreeMap<String, Float> = TreeMap())
+        val rates: TreeMap<String, BigDecimal> = TreeMap())
